# Conflicts:
#	src/main/java/supersymmetry/api/recipes/SuSyRecipeMaps.java
#	src/main/java/supersymmetry/api/unification/ore/SusyOrePrefix.java
#	src/main/java/supersymmetry/common/item/SuSyMetaItems.java
#	src/main/java/supersymmetry/common/metatileentities/SuSyMetaTileEntities.java
#	src/main/java/supersymmetry/common/metatileentities/multi/electric/MetaTileEntityBallMill.java
#	src/main/java/supersymmetry/loaders/recipes/SusyOreRecipeHandler.java
#	src/main/resources/assets/susy/lang/en_us.lang

+++ src/main/java/supersymmetry/loaders/recipes/SusyOreRecipeHandler.java
@@ -1,5 +1,7 @@
 package supersymmetry.loaders.recipes;
 
+import net.minecraft.item.ItemStack;
+
 import gregtech.api.recipes.ModHandler;
 import gregtech.api.recipes.RecipeBuilder;
 import gregtech.api.recipes.RecipeMaps;
@@ -13,9 +15,6 @@ import gregtech.api.unification.stack.MaterialStack;
 import gregtech.api.unification.stack.UnificationEntry;
 import gregtech.api.util.GTUtility;
 import gregtech.common.ConfigHolder;
-import gregtech.loaders.recipe.handlers.OreRecipeHandler;
-import net.minecraft.item.ItemStack;
-import supersymmetry.api.recipes.SuSyRecipeMaps;
 import supersymmetry.api.unification.ore.SusyOrePrefix;
 
 public class SusyOreRecipeHandler {
@@ -59,8 +58,9 @@ public class SusyOreRecipeHandler {
         }
 
         ItemStack crushedStack = OreDictUnifier.get(OrePrefix.crushed, material);
-        Material smeltingMaterial = property.getDirectSmeltResult() == null ? material : property.getDirectSmeltResult();
-        double amountOfCrushedOre = property.getOreMultiplier();
+        Material smeltingMaterial = property.getDirectSmeltResult() == null ? material :
+                property.getDirectSmeltResult();
+        double amountOfCrushedOre = (double) property.getOreMultiplier();
         ItemStack ingotStack;
         if (smeltingMaterial.hasProperty(PropertyKey.INGOT)) {
             ingotStack = OreDictUnifier.get(OrePrefix.ingot, smeltingMaterial);
@@ -84,7 +84,7 @@ public class SusyOreRecipeHandler {
             builder.buildAndRegister();
             builder = RecipeMaps.MACERATOR_RECIPES.recipeBuilder().input(orePrefix, material).outputs(GTUtility.copy((int)Math.round(amountOfCrushedOre) * 2 * oreTypeMultiplier, crushedStack)).chancedOutput(byproductStack, 1400, 850).duration(400);
 
-            for(MaterialStack secondaryMaterial : orePrefix.secondaryMaterials) {
+            for (MaterialStack secondaryMaterial : orePrefix.secondaryMaterials) {
                 if (secondaryMaterial.material.hasProperty(PropertyKey.DUST)) {
                     ItemStack dustStack = OreDictUnifier.getGem(secondaryMaterial);
                     builder.chancedOutput(dustStack, 6700, 800);
@@ -136,5 +136,4 @@ public class SusyOreRecipeHandler {
     private static boolean doesMaterialUseNormalFurnace(Material material) {
         return !material.hasProperty(PropertyKey.BLAST);
     }
-
 }
